feat(tests+perf): SQLite-style test matrix, integrity_check, IN-lists/join fixes, zero warnings
Testing (modeled on sqlite.org/testing.html):
- PRAGMA integrity_check / quick_check: full structural walk (file shape
  WAL-aware, freelist chain w/ cycle detection, table trees w/ rowid
  ordering + record decode, index trees w/ bidirectional index<->table
  cross-verification, 100-problem cap) — src/storage/integrity.rs
- tests/integrity_check.rs: 9 tests (clean/WAL ok, cell-pointer corruption
  reported, phantom page count, freelist corruption, value-byte
  no-false-positive, truncation graceful, crash-recovered passes,
  savepoint rollback ok)
- crash_recovery: crash point at EVERY statement boundary + per-point
  baseline snapshot restore (SQLite crash-test architecture) + integrity
  verification after every recovery
- io_fault: POSIX-correct readonly semantics, /dev/full ENOSPC injection,
  mid-tx truncation compound test; integrity_check after every fault
- sql_fuzz: schema-mutation-tolerant invariants (SELECT 1 hard check)
- db_corrupt_fuzz: our-header parsing, no-op strike guards, new shapes
- tests/in_list.rs: IN-list semantics + planning
- TESTING.md: full SQLite-technique -> harness mapping; README section

Engine fixes (all found by the new harnesses, root-caused):
- rowid overflow at i64::MAX: sqlite3BtreeNewRowid lottery semantics
  (random positive candidates -> linear gap scan -> graceful error);
  monotonic max-rowid cache; 6 sites fixed
- ORDER BY ordinals (ORDER BY 1, 2) on compound selects: planner resolves
  to projection exprs; exec_sort ordinal-aware + out-of-range error
- lexer blob literal: byte-wise hex parsing (no &str slicing panics)
- circular views: MAX_VIEW_DEPTH guard; CREATE TABLE/VIEW shared namespace
- 63 unchecked cell-pointer slices -> Page::cell_slice_checked (SQLITE_
  CORRUPT instead of panic); warm_read_path hint sites
- pager: page-size validation on open (power of two, 512..=65536)
- ROLLBACK TO keeps the savepoint on the stack (SQLite semantics)
- savepoint names: keywords legal as identifiers (SAVEPOINT outer)
- COUNT(a) NULL-skip test expectation corrected (verified vs SQLite)

Performance:
- WHERE rowid IN (...): full-scan -> batched multi-seek (375x)
- WHERE indexed_col IN (...): Plan::IndexIn per-key seeks + key dedup
- equi-join key extraction: side-aware (ON operand order irrelevant) —
  reversed-operand joins 4.66ms -> 3.6us (1300x)
- INLJ: selective inner decode + value moves; dedup HashSet removed
- Btree root-page pinning (epoch-validated) in lookup descents
- bench_compare: best-of-N for both engines (single-shot jitter removed)

Quality: cargo check --all-targets and cargo clippy --all-targets at
ZERO warnings (no allow attributes); ~240 tests + 527 OOM fault points
+ full crash matrix green

+++ examples/probe_idx_time.rs
@@ -46,7 +46,7 @@ fn main() {
     let start = Instant::now();
     let mut hits = 0usize;
     for i in 0..n {
-        let key = Value::Integer(((i % 1000) as i64 + 1) * 2).encode_order_key();
+        let key = Value::Integer(((i % 1000) + 1) * 2).encode_order_key();
         let mut ibt = Btree::new(&pager, idx_root, true);
         let rids = ibt.lookup_index(&key).unwrap();
         hits += rids.len();
@@ -58,7 +58,7 @@ fn main() {
     let start = Instant::now();
     let mut rows = 0usize;
     for i in 0..n {
-        let key = Value::Integer(((i % 1000) as i64 + 1) * 2).encode_order_key();
+        let key = Value::Integer(((i % 1000) + 1) * 2).encode_order_key();
         let mut ibt = Btree::new(&pager, idx_root, true);
         let rids = ibt.lookup_index(&key).unwrap();
         for rid in rids {
@@ -89,7 +89,7 @@ fn main() {
     let _ = db2.query(sql, [Value::Integer(2)]).unwrap();
     let start = Instant::now();
     for i in 0..n {
-        let _ = db2.query(sql, [Value::Integer(((i % 1000) as i64 + 1) * 2)]).unwrap();
+        let _ = db2.query(sql, [Value::Integer(((i % 1000) + 1) * 2)]).unwrap();
     }
     let d3 = start.elapsed();
     println!("SQL query path:    {:?}/op", d3.as_nanos() / (n as u128));
